+++ src/solver/optim-model-filler/include/antares/solver/optim-model-filler/ReadLinearExpressionVisitor.h
@@ -62,11 +62,11 @@ public:
     {
     }
 
-    Antares::Optimization::TimeDependentLinearExpression visitRemoveDuplicates(
+    Antares::Optimization::TimeDependentLinearExpression visitMergeDuplicates(
       const Nodes::Node* node)
     {
         auto expr = dispatch(node);
-        expr.removeDuplicateCoefficients();
+        expr.mergeDuplicateCoefficients();
         return expr;
     }
